fix(tests): avoid ambiguous calls

+++ tests/sources/test_constants.cpp
@@ -33,28 +33,28 @@ using namespace rtm;
 TEST_CASE("constants", "[math][constants]")
 {
 	// Float
-	CHECK(scalar_near_equal(constants::pi(), 3.141592653589793238462643383279502884F, 1.0e-6F));
-	CHECK(scalar_near_equal(-constants::pi(), -3.141592653589793238462643383279502884F, 1.0e-6F));
-	CHECK(scalar_near_equal(+constants::pi(), +3.141592653589793238462643383279502884F, 1.0e-6F));
-	CHECK(scalar_near_equal(constants::pi() * 2.0F, 3.141592653589793238462643383279502884F * 2.0F, 1.0e-6F));
-	CHECK(scalar_near_equal(2.0F * constants::pi(), 2.0F * 3.141592653589793238462643383279502884F, 1.0e-6F));
-	CHECK(scalar_near_equal(constants::pi() / 2.0F, 3.141592653589793238462643383279502884F / 2.0F, 1.0e-6F));
-	CHECK(scalar_near_equal(2.0F / constants::pi(), 2.0F / 3.141592653589793238462643383279502884F, 1.0e-6F));
-	CHECK(scalar_near_equal(constants::pi() + 1.0F, 3.141592653589793238462643383279502884F + 1.0F, 1.0e-6F));
-	CHECK(scalar_near_equal(1.0F + constants::pi(), 1.0F + 3.141592653589793238462643383279502884F, 1.0e-6F));
-	CHECK(scalar_near_equal(constants::pi() - 1.0F, 3.141592653589793238462643383279502884F - 1.0F, 1.0e-6F));
-	CHECK(scalar_near_equal(1.0F - constants::pi(), 1.0F - 3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(constants::pi()), 3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(-constants::pi()), -3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(+constants::pi()), +3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(constants::pi()) * 2.0F, 3.141592653589793238462643383279502884F * 2.0F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(2.0F * constants::pi()), 2.0F * 3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(constants::pi()) / 2.0F, 3.141592653589793238462643383279502884F / 2.0F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(2.0F / constants::pi()), 2.0F / 3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(constants::pi()) + 1.0F, 3.141592653589793238462643383279502884F + 1.0F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(1.0F + constants::pi()), 1.0F + 3.141592653589793238462643383279502884F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(constants::pi()) - 1.0F, 3.141592653589793238462643383279502884F - 1.0F, 1.0e-6F));
+	CHECK(scalar_near_equal(float(1.0F - constants::pi()), 1.0F - 3.141592653589793238462643383279502884F, 1.0e-6F));
 
 	// Double
-	CHECK(scalar_near_equal(constants::pi(), 3.141592653589793238462643383279502884, 1.0e-6));
-	CHECK(scalar_near_equal(-constants::pi(), -3.141592653589793238462643383279502884, 1.0e-6));
-	CHECK(scalar_near_equal(+constants::pi(), +3.141592653589793238462643383279502884, 1.0e-6));
-	CHECK(scalar_near_equal(constants::pi() * 2.0, 3.141592653589793238462643383279502884 * 2.0, 1.0e-6));
-	CHECK(scalar_near_equal(2.0 * constants::pi(), 2.0 * 3.141592653589793238462643383279502884, 1.0e-6));
-	CHECK(scalar_near_equal(constants::pi() / 2.0, 3.141592653589793238462643383279502884 / 2.0, 1.0e-6));
-	CHECK(scalar_near_equal(2.0 / constants::pi(), 2.0 / 3.141592653589793238462643383279502884, 1.0e-6));
-	CHECK(scalar_near_equal(constants::pi() + 1.0, 3.141592653589793238462643383279502884 + 1.0, 1.0e-6));
-	CHECK(scalar_near_equal(1.0 + constants::pi(), 1.0 + 3.141592653589793238462643383279502884, 1.0e-6));
-	CHECK(scalar_near_equal(constants::pi() - 1.0, 3.141592653589793238462643383279502884 - 1.0, 1.0e-6));
-	CHECK(scalar_near_equal(1.0 - constants::pi(), 1.0 - 3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(constants::pi()), 3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(-constants::pi()), -3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(+constants::pi()), +3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(constants::pi() * 2.0), 3.141592653589793238462643383279502884 * 2.0, 1.0e-6));
+	CHECK(scalar_near_equal(double(2.0 * constants::pi()), 2.0 * 3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(constants::pi() / 2.0), 3.141592653589793238462643383279502884 / 2.0, 1.0e-6));
+	CHECK(scalar_near_equal(double(2.0 / constants::pi()), 2.0 / 3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(constants::pi() + 1.0), 3.141592653589793238462643383279502884 + 1.0, 1.0e-6));
+	CHECK(scalar_near_equal(double(1.0 + constants::pi()), 1.0 + 3.141592653589793238462643383279502884, 1.0e-6));
+	CHECK(scalar_near_equal(double(constants::pi() - 1.0), 3.141592653589793238462643383279502884 - 1.0, 1.0e-6));
+	CHECK(scalar_near_equal(double(1.0 - constants::pi()), 1.0 - 3.141592653589793238462643383279502884, 1.0e-6));
 }
